refactor(storage): extract memory_bundles to its own module (R.3)
Phase R.3 of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Smallest, lowest-risk
subsystem extraction — moves the Memory bundle struct + the 4
bundle_memory_* CRUD methods + the map_memory_row helper out of
store.rs (6226 lines) into a new memory_bundles.rs (151 lines).

Method surface is unchanged: `Store::bundle_memory_*` still lives
on `Store` thanks to a sibling `impl Store {}` block in the new
module. The `Memory` struct is re-exported from `storage::store`
AND from `storage::` so existing imports
(`storage::store::Memory`, `storage::Memory`) keep working without
churn at call sites.

Made `Store::conn` `pub(super)` so sibling subsystem modules can
take the connection lock. This is a one-time visibility bump for
all R.x extractions — future R.4 / R.5 / R.6 will reuse it.

Net store.rs change: -227 lines (mostly the moved blocks; some
blank-line consolidation). New file 151 lines.

154/154 storage tests + bundle_memory lifecycle test green.
