+++ .changesets/1779970122-refactor-storage-extract-memory-bundles-to-its-own-module-r--4ugk.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(storage): extract memory_bundles to its own module (R.3)
